initial commit rename main controller add popover
